booking-fastapi-backend/pyproject.toml
[project]
name = "booking-fastapi-app"
version = "0.1.0"
description = "A FastAPI booking application."
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "pydantic",
    "sqlalchemy",
    "passlib[argon2]",
    "pyjwt",
    "pymysql",
    "cryptography",
]

[build-system]
requires = ["setuptools>=61.0.0"]
build-backend = "setuptools.build_meta"

# =======================================================
# FIX: Explicitly list modules for setuptools to resolve
# the "Multiple top-level modules" error during editable install.
# =======================================================
[tool.setuptools]
# Only the app package is needed
packages = ["app"]

# =======================================================
# Code Quality and Formatting Tools
# =======================================================

# Black - Code Formatter
[tool.black]
line-length = 100
target-version = ['py311']
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

# isort - Import Sorter (Black compatible)
[tool.isort]
profile = "black"
line_length = 100
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

# Ruff - Fast Python Linter
[tool.ruff]
line-length = 100
target-version = "py311"
exclude = [
    ".git",
    "__pycache__",
    ".venv",
    "venv",
]

[tool.ruff.lint]
select = [
    "E",      # PyCodeStyle errors
    "W",      # PyCodeStyle warnings
    "F",      # PyFlakes
    "I",      # isort
    "N",      # pep8-naming
    "UP",     # pyupgrade
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "SIM",    # flake8-simplify
]
ignore = [
    "E501",   # line too long (handled by Black)
    "N999",   # invalid module name (for hyphenated packages)
    "B008",   # do not perform function call in argument defaults (FastAPI Depends pattern)
]

# Pytest Configuration
[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
addopts = "--strict-markers"

# Mypy Configuration (Optional Type Checking)
[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true
